⚡ chore: move events to /src
