+++ src/app/index.ts
@@ -1,2 +1,1 @@
-export { events } from './events';
 export { config } from './config';
