+++ src/components/ui/primitives/compound/welcome.tsx
@@ -1,4 +1,4 @@
-import { events } from 'app';
+import { events } from '@events';
 
 import { Text } from 'components/ui/primitives/atoms/text';
 import { ShareModal } from 'components/ui/common/modals/share';
