+++ src/features/image/index.ts
@@ -3,7 +3,7 @@ import dynamic from 'next/dynamic';
 import { imageSectionParser } from './parser';
 import { defaultImageSectionConfig } from './default-config';
 
-import { events } from 'app';
+import { events } from '@events';
 import { PanelsEnum, Sections } from 'types';
 
 const feature = {
