+++ src/features/stats/panel/views/layout/item/index.tsx
@@ -5,7 +5,7 @@ import { Tile } from 'components/ui/primitives/atoms/tile';
 import { Icon } from 'components/ui/primitives/atoms/icon';
 import { Tooltip } from 'components/ui/primitives/atoms/tooltip';
 
-import { events } from 'app';
+import { events } from '@events';
 import { variants, animations } from './animations';
 
 type ItemProps = {
