+++ src/features/techs/panel/views/adding/index.tsx
@@ -5,7 +5,7 @@ import { Text } from 'components/ui/primitives/atoms/text';
 import { Panel } from 'components/ui/primitives/atoms/panel';
 import { DisplayBlock } from 'components/ui/primitives/atoms/display-block';
 
-import { events } from 'app';
+import { events } from '@events';
 import { useForceUpdate } from 'hooks';
 import { capitalize, cn, debounce, filterArrayByQueryMatch } from 'utils';
