+++ src/features/techs/panel/views/editing/providers/index.tsx
@@ -1,7 +1,7 @@
 import { Text } from 'components/ui/primitives/atoms/text';
 import { DropdownMenu } from 'components/ui/primitives/atoms/dropdown-menu';
 
-import { events } from 'app';
+import { events } from '@events';
 import { IconProviders } from 'types';
 
 type ProvidersProps = {
