+++ src/features/text/index.ts
@@ -4,7 +4,7 @@ import dynamic from 'next/dynamic';
 import { textSectionParser } from './parser';
 import { defaultTextSectionConfig } from './default-config';
 
-import { events } from 'app';
+import { events } from '@events';
 import { PanelsEnum, Sections } from 'types';
 
 const feature = {
